Fix circular import for PipelineComposerRequirements
