+++ fedot/core/composer/gp_composer/gp_composer.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from typing import Collection, Optional, Sequence, Tuple, Union
 
 from fedot.core.caching.pipelines_cache import OperationsCache
@@ -6,46 +5,12 @@ from fedot.core.caching.preprocessing_cache import PreprocessingCache
 from fedot.core.composer.composer import Composer
 from fedot.core.data.data import InputData
 from fedot.core.data.multi_modal import MultiModalData
-from fedot.core.optimisers.gp_comp.operators.mutation import MutationStrengthEnum
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.graph import OptGraph
 from fedot.core.optimisers.objective.data_objective_builder import DataObjectiveBuilder
 from fedot.core.optimisers.opt_history import OptHistory
 from fedot.core.optimisers.optimizer import GraphOptimizer
-from fedot.core.optimisers.composer_requirements import ComposerRequirements
 from fedot.core.pipelines.pipeline import Pipeline
-
-
-@dataclass
-class PipelineComposerRequirements(ComposerRequirements):
-    """
-    Dataclass is for defining the requirements for composition process of genetic programming composer
-
-    :attribute pop_size: initial population size; if unspecified, default value is used.
-    :attribute max_pop_size: maximum population size; optional, if unspecified, then population size is unbound.
-    :attribute num_of_generations: maximal number of evolutionary algorithm generations
-    :attribute crossover_prob: crossover probability (the chance that two chromosomes exchange some of their parts)
-    :attribute mutation_prob: mutation probability
-    :attribute mutation_strength: strength of mutation in tree (using in certain mutation types)
-    :attribute max_pipeline_fit_time: time constraint for operation fitting (minutes)
-    :attribute start_depth: start value of tree depth
-    :attribute validation_blocks: number of validation blocks for time series validation
-    :attribute n_jobs: num of n_jobs
-    :attribute collect_intermediate_metric: save metrics for intermediate (non-root) nodes in pipeline
-    :attribute keep_n_best: Number of the best individuals of previous generation to keep in next generation.
-    """
-    pop_size: int = 20
-    max_pop_size: Optional[int] = 55
-    num_of_generations: int = 20
-    offspring_rate: float = 0.5
-    crossover_prob: float = 0.8
-    mutation_prob: float = 0.8
-    mutation_strength: MutationStrengthEnum = MutationStrengthEnum.mean
-    max_pipeline_fit_time: int = None
-    start_depth: int = None
-    validation_blocks: int = None
-    n_jobs: int = 1
-    collect_intermediate_metric: bool = False
-    keep_n_best: int = 1
 
 
 class GPComposer(Composer):
