+++ fedot/core/optimisers/gp_comp/operators/inheritance.py
@@ -1,7 +1,7 @@
 from functools import partial
 from typing import (Callable)
 
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT, Operator
 from fedot.core.optimisers.gp_comp.operators.selection import Selection
 from fedot.core.utilities.data_structures import ComparableEnum as Enum
