+++ fedot/core/optimisers/gp_comp/operators/selection.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 from random import choice, randint
 from typing import List, Callable
 
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT, Operator
 from fedot.core.utilities.data_structures import ComparableEnum as Enum
