+++ fedot/core/optimisers/gp_comp/parameters/operators_prob.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.gp_comp.operators.inheritance import GeneticSchemeTypesEnum
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT
 from fedot.core.optimisers.gp_comp.parameters.mutation_prob import AdaptiveMutationProb
