+++ fedot/core/optimisers/initial_graphs_generator.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional, Sequence, Union, Iterable
 
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.constants import MAXIMAL_ATTEMPTS_NUMBER
 from fedot.core.dag.graph import Graph
 from fedot.core.log import default_log
