+++ test/unit/composer/test_history.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 from fedot.api.main import Fedot
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.dag.graph import Graph
 from fedot.core.dag.verification_rules import DEFAULT_DAG_RULES
 from fedot.core.data.data import InputData
