+++ test/unit/optimizer/test_default_node_factory.py
@@ -1,5 +1,5 @@
 from fedot.core.composer.advisor import PipelineChangeAdvisor
-from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
+from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.graph import OptNode
 from fedot.core.optimisers.opt_node_factory import DefaultOptNodeFactory
 from fedot.core.repository.tasks import Task, TaskTypesEnum
